fix extraData function reference
